+++ crates/crypto/pqc/src/hashsig/mod.rs
@@ -1,11 +1,4 @@
-pub mod keystore;
+pub mod errors;
 pub mod private_key;
 pub mod public_key;
 pub mod signature;
-
-#[cfg(test)]
-mod tests;
-
-pub use private_key::PrivateKey;
-pub use public_key::PublicKey;
-pub use signature::Signature;
