+++ crates/crypto/pqc/src/traits.rs
@@ -1,12 +1,14 @@
-use crate::hashsig::{Signature, public_key::PublicKey};
+use crate::hashsig::{public_key::PublicKey, signature::Signature};
 
 pub trait PQSignable {
     type Error;
+
     fn sign(&self, message: &[u8], epoch: u32) -> Result<Signature, Self::Error>;
 }
 
 pub trait PQVerifiable {
     type Error;
+
     fn verify(
         &self,
         message: &[u8],
